优化了超声接收和信号处理部分代码

+++ Ultrasonic_Vehicle_2025/src/ultrasonic_controller.h
@@ -30,7 +30,7 @@ public:
     bool available();
 
     // 获取左右到达时间差（微秒）
-    long getTimeDiff();
+    int16_t getTimeDiff();
 
     // 手动复位（一般不需要外部调用）
     void reset();
